asArray replaced
